challenge6/requirements.txt
python-dotenv
openai
azure-monitor-opentelemetry
opentelemetry-api
azure-cosmos